requirements.txt
scipy
clean_fid==0.1.35
comet_ml
imageio
numpy
pandas
torch
scikit_learn
torchvision
tqdm==4.64.0
wandb
# faiss-gpu
faiss-gpu-cu12
datasets
pillow
